moved sasconfig to config ; fixed related tests and setup.py issues; passed all tests
