+++ setup.py
@@ -53,7 +53,7 @@ setup(name='sasmol',
 
 	package_dir={'sasmol':os.path.join('src','python')},
 
-    packages=['sasmol','sasmol.test_sasmol','sasmol.test_sasmol.util','sasmol.test_sasmol.manual_tests','sasmol.test_sasmol.data','sasmol.test_sasmol.data.pdb_common','sasmol.test_sasmol.data.dcd_common','sasmol.test_sasmol.data.sasmol','sasmol.test_sasmol.data.sasmol.calculate','sasmol.test_sasmol.data.sasmol.sasio','sasmol.test_sasmol.data.sasmol.sasio.test-results','sasmol.test_sasmol.data.sasmol.sasmath','sasmol.test_sasmol.data.sasmol.sasmol','sasmol.test_sasmol.data.sasmol.operate','sasmol.test_sasmol.data.sasmol.sasproperties','sasmol.test_sasmol.test_calculate','sasmol.test_sasmol.test_sasio','sasmol.test_sasmol.test_sasmath','sasmol.test_sasmol.test_sasmol','sasmol.test_sasmol.test_operate','sasmol.test_sasmol.test_sasproperties','sasmol.test_sasmol.test_sassubset','sasmol.extensions','sasmol.extensions.dcdio','sasmol.extensions.view','sasmol.extensions.mask','sasmol.extensions.matrix_math'],
+    packages=['sasmol','sasmol.test_sasmol','sasmol.test_sasmol.util','sasmol.test_sasmol.manual_tests','sasmol.test_sasmol.data','sasmol.test_sasmol.data.pdb_common','sasmol.test_sasmol.data.dcd_common','sasmol.test_sasmol.data.sasmol','sasmol.test_sasmol.data.sasmol.calculate','sasmol.test_sasmol.data.sasmol.sasio','sasmol.test_sasmol.data.sasmol.sasio.test-results','sasmol.test_sasmol.data.sasmol.sasmath','sasmol.test_sasmol.data.sasmol.sasmol','sasmol.test_sasmol.data.sasmol.operate','sasmol.test_sasmol.data.sasmol.properties','sasmol.test_sasmol.test_calculate','sasmol.test_sasmol.test_sasio','sasmol.test_sasmol.test_sasmath','sasmol.test_sasmol.test_sasmol','sasmol.test_sasmol.test_operate','sasmol.test_sasmol.test_properties','sasmol.test_sasmol.test_sassubset','sasmol.extensions','sasmol.extensions.dcdio','sasmol.extensions.view','sasmol.extensions.mask','sasmol.extensions.matrix_math'],
 	
     ext_modules=[
 	Extension('sasmol._dcdio',[os.path.join('src','python','extensions','dcdio','dcdio.i'),os.path.join('src','python','extensions','dcdio','dcdio.c')],include_dirs=[numpy_include]),
@@ -91,18 +91,18 @@ setup(name='sasmol',
         ( os.path.join('sasmol','test_sasmol','data','sasmol','calculate') , [os.path.join('src','python','test_sasmol','data','sasmol','calculate','1CRN-rot-shift.pdb')]),
     
     
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','Catoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','ConflictAtoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','Hatoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','MisAtoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','Natoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','Oatoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','Otheratoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','Patoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','Satoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','amino_acid_sld.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','charmm27_atoms.txt')]),
-        ( os.path.join('sasmol','test_sasmol','data','sasmol','sasproperties') , [os.path.join('src','python','test_sasmol','data','sasmol','sasproperties','standard_atomic_weigh.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','Catoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','ConflictAtoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','Hatoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','MisAtoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','Natoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','Oatoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','Otheratoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','Patoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','Satoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','amino_acid_sld.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','charmm27_atoms.txt')]),
+        ( os.path.join('sasmol','test_sasmol','data','sasmol','properties') , [os.path.join('src','python','test_sasmol','data','sasmol','properties','standard_atomic_weigh.txt')]),
    
         ( os.path.join('sasmol','test_sasmol','data','sasmol','operate') , [os.path.join('src','python','test_sasmol','data','sasmol','operate','1CRN-rot-shift.pdb')]),
         ( os.path.join('sasmol','test_sasmol','data','sasmol','operate') , [os.path.join('src','python','test_sasmol','data','sasmol','operate','1CRN-rot-sub.pdb')]),
