+++ src/python/__init__.py
@@ -1,4 +1,4 @@
-import sasconfig as sasconfig
+import config as config
 
 #global __sasmol_id__ 
 #__sasmol_id__ = 0
@@ -10,10 +10,10 @@ import sasconfig as sasconfig
 #__sasmol_logging__ = False
 #
 #
-#if sasconfig.__level__ == 'DEBUG':
+#if config.__level__ == 'DEBUG':
 #    __sasmol_debug__ = False
 #
-#if sasconfig.__logging__ == True:
+#if config.__logging__ == True:
 #    __sasmol_logging__ = True
 #
