+++ src/python/test_sasmol/test_calculate/test_intg_sascalc_Prop_calc_minmax_all_steps.py
@@ -14,9 +14,7 @@
     You should have received a copy of the GNU General Public License
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 '''
-
 from sasmol.test_sasmol.util import env, util
-
 from unittest import main, skipIf
 from mocker import Mocker, MockerTestCase, ANY, ARGS, KWARGS
 import sasmol.sasmol as sasmol
