+++ src/python/test_sasmol/test_calculate/test_unit_sascalc_Prop_calcrmsd.py
@@ -15,6 +15,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 '''
 
+import sasmol.test_sasmol.util as util
 from sasmol.test_sasmol.util import env, util
 
 from unittest import main
